validator/requirements.txt
flask
pika
pymongo
requests
mysql-connector-python
